requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-extensions==2.0.6
django-rest-framework==0.1.0
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pytz==2018.3
requests==2.18.4
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22